refactor: FoodCategory 반환 형식 수정
value를 추가

Related to: #115
