+++ backend/src/main/java/com/multi/mariage/category/vo/food/FoodCategoriesVO.java
@@ -11,17 +11,20 @@ import lombok.NoArgsConstructor;
 public class FoodCategoriesVO {
     private int id;
     private String name;
+    private FoodCategory value;
 
     @Builder
-    private FoodCategoriesVO(int id, String name) {
+    private FoodCategoriesVO(int id, String name, FoodCategory value) {
         this.id = id;
         this.name = name;
+        this.value = value;
     }
 
     public static FoodCategoriesVO from(FoodCategory category) {
         return FoodCategoriesVO.builder()
                 .id(category.getId())
                 .name(category.getName())
+                .value(category)
                 .build();
     }
 }
