+++ backend/src/main/java/com/multi/mariage/product/service/ProductService.java
@@ -2,7 +2,7 @@ package com.multi.mariage.product.service;
 
 import com.multi.mariage.product.domain.Product;
 import com.multi.mariage.product.domain.ProductRepository;
-import com.multi.mariage.product.dto.ProductSaveRequest;
+import com.multi.mariage.product.dto.request.ProductSaveRequest;
 import com.multi.mariage.storage.domain.Image;
 import com.multi.mariage.storage.repository.StorageRepository;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +20,7 @@ public class ProductService {
 
     @Transactional
     public Product create(ProductSaveRequest request) {
-//TODO: 현재 임시로 예외처리. 예외처리 및 검증 구조 따로 생성 필요.
+        //TODO: 현재 임시로 예외처리. 예외처리 및 검증 구조 따로 생성 필요.
         log.info("");
         Image image = storageRepository.findById(request.getImageId())
                 .orElseThrow(() -> new IllegalArgumentException("이미지를 찾을 수 없습니다."));
@@ -29,7 +29,7 @@ public class ProductService {
                 .name(request.getName())
                 .level(request.getLevel())
                 .info(request.getInfo())
-                .country(request.getCountry())
+//                .country(request.getCountry())
                 .image(image)
                 .build();
         product.setImage(image);
